grandTotal issue in Decimal point

+++ src/main/java/com/asg/finance/dto/ApPurchaseInvoiceItemDtlDto.java
@@ -3,6 +3,7 @@ package com.asg.finance.dto;
 import com.asg.common.lib.dto.LovGetListDto;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -16,9 +17,9 @@ public class ApPurchaseInvoiceItemDtlDto {
     private Long poQty;
     private Long dnQty;
     private Long qtyReceived;
-    private Long price;
+    private BigDecimal price;
     private Long discount;
-    private Long total;
+    private BigDecimal total;
     private String remarks;
     private String createdBy;
     private LocalDateTime createdDate;
@@ -30,9 +31,9 @@ public class ApPurchaseInvoiceItemDtlDto {
     private Long refDetRowId;
     private Long taxPoid;
     private Long taxPercentage;
-    private Long taxAmount;
-    private Long amount;
-    private Long baseAmount;
+    private BigDecimal taxAmount;
+    private BigDecimal amount;
+    private BigDecimal baseAmount;
 
     private LovGetListDto stockDet;
     private LovGetListDto stockUnitDet;
